requirements.txt
Django==1.11.20
django-contrib-comments==1.9.1
django-filter==2.1.0
requests==2.21.0
requests-oauthlib==1.0.0
Mezzanine==4.3.1
Cartridge==0.13.0
gunicorn==19.9.0
psycopg2==2.7.7
whitenoise==4.1.2
dj-database-url==0.5.0
django-widget-tweaks==1.4.3
stripe==1.84.2
django-clear-cache==0.3
python-memcached==1.59
boto==2.49.0
boto3==1.9.134
django-storages==1.7.1
mezzanine_heroku==0.3.25
mezzanine-cartridge-api==1.5.17
newrelic==4.14.0.115
